tg_bot/requirements.txt
aiohttp~=3.11.13
aiogram~=3.18.0
python-dotenv~=1.0.1